+++ src/bundle/Resources/translations/change_password.en.xliff
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<xliff xmlns="urn:oasis:names:tc:xliff:document:1.2" xmlns:jms="urn:jms:translation" version="1.2">
+  <file source-language="en" target-language="en" datatype="plaintext" original="not.available">
+    <header>
+      <tool tool-id="JMSTranslationBundle" tool-name="JMSTranslationBundle" tool-version="1.1.0-DEV"/>
+      <note>The source node in most cases contains the sample message as written by the developer. If it looks like a dot-delimitted string such as "form.label.firstname", then the developer has not provided a default message.</note>
+    </header>
+    <body>
+      <trans-unit id="014a9174f404ffba404748439bd1fa00acf16113" resname="ezplatform.change_password.success">
+        <source>Your password has been successfully changed.</source>
+        <target state="new">Your password has been successfully changed.</target>
+        <note>key: ezplatform.change_password.success</note>
+      </trans-unit>
+    </body>
+  </file>
+</xliff>
