+++ src/bundle/Resources/translations/forgot_password.en.xliff
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<xliff xmlns="urn:oasis:names:tc:xliff:document:1.2" xmlns:jms="urn:jms:translation" version="1.2">
+  <file source-language="en" target-language="en" datatype="plaintext" original="not.available">
+    <header>
+      <tool tool-id="JMSTranslationBundle" tool-name="JMSTranslationBundle" tool-version="1.1.0-DEV"/>
+      <note>The source node in most cases contains the sample message as written by the developer. If it looks like a dot-delimitted string such as "form.label.firstname", then the developer has not provided a default message.</note>
+    </header>
+    <body>
+      <trans-unit id="aa927d2f4bf95b09884f3f373c05cb0a268d0147" resname="ezplatform.forgot_password.contact_administrator">
+        <source>If you do not remember your login, please contact the site administrator.</source>
+        <target state="new">If you do not remember your login, please contact the site administrator.</target>
+        <note>key: ezplatform.forgot_password.contact_administrator</note>
+      </trans-unit>
+      <trans-unit id="b88c08f07a426c1fb80c9e94e4d2fca7bc0832f2" resname="ezplatform.forgot_password.login">
+        <source>This email is connected with several accounts. Please enter your login.</source>
+        <target state="new">This email is connected with several accounts. Please enter your login.</target>
+        <note>key: ezplatform.forgot_password.login</note>
+      </trans-unit>
+      <trans-unit id="1af697cb32b6eeeafbf4ca827e766829eac0e992" resname="ezplatform.forgot_password.message">
+        <source><![CDATA[Hi,
+        <br /><br />
+        We have received a request to reset the password for your account on site. Click “reset password” below to choose a new password:
+        <br /><br />
+        <a href="%reset_password%">Reset password<a/>
+        <br /><br />
+        If you did not request a password reset, please ignore this email, and your password will remain the same.]]></source>
+        <target state="new"><![CDATA[Hi,
+        <br /><br />
+        We have received a request to reset the password for your account on site. Click “reset password” below to choose a new password:
+        <br /><br />
+        <a href="%reset_password%">Reset password<a/>
+        <br /><br />
+        If you did not request a password reset, please ignore this email, and your password will remain the same.]]></target>
+        <note>key: ezplatform.forgot_password.message</note>
+      </trans-unit>
+      <trans-unit id="9d77404b1f40b0458d3b82757b067e15c7dc46e3" resname="ezplatform.forgot_password.reset_your_password">
+        <source>Reset your password</source>
+        <target state="new">Reset your password</target>
+        <note>key: ezplatform.forgot_password.reset_your_password</note>
+      </trans-unit>
+      <trans-unit id="9b35c814c4edde4e4ecc46eef81c1be1821601ba" resname="ezplatform.forgot_password.success">
+        <source><![CDATA[<p>We have sent an email to the email address of your account. You should receive the email shortly.</p>
+            <p>If the email is slow to arrive, check your spam folders.</p>
+            <p>If you receive no email, it’s that your username or email is unknown to us, so please double check.</p>
+            <p>If you reset your password multiple times, only the most recent email link will be valid</p>]]></source>
+        <target state="new"><![CDATA[<p>We have sent an email to the email address of your account. You should receive the email shortly.</p>
+            <p>If the email is slow to arrive, check your spam folders.</p>
+            <p>If you receive no email, it’s that your username or email is unknown to us, so please double check.</p>
+            <p>If you reset your password multiple times, only the most recent email link will be valid</p>]]></target>
+        <note>key: ezplatform.forgot_password.success</note>
+      </trans-unit>
+    </body>
+  </file>
+</xliff>
