+++ src/lib/Form/Data/UserPasswordChangeData.php
@@ -6,12 +6,21 @@
  */
 declare(strict_types=1);
 
-namespace EzSystems\EzPlatformUser\Form\Data\User;
+namespace EzSystems\EzPlatformUser\Form\Data;
 
 use Symfony\Component\Validator\Constraints as Assert;
+use EzSystems\EzPlatformUser\Validator\Constraints as UserAssert;
 
-class UserPasswordResetData
+class UserPasswordChangeData
 {
+    /**
+     * @UserAssert\UserPassword()
+     * @Assert\NotBlank()
+     *
+     * @var string
+     */
+    private $oldPassword;
+
     /**
      * @Assert\NotBlank()
      *
@@ -20,11 +29,21 @@ class UserPasswordResetData
     private $newPassword;
 
     /**
+     * @param string|null $oldPassword
      * @param string|null $newPassword
      */
-    public function __construct(?string $newPassword = null)
+    public function __construct(?string $oldPassword = null, ?string $newPassword = null)
     {
+        $this->oldPassword = $oldPassword;
         $this->newPassword = $newPassword;
+    }
+
+    /**
+     * @param string|null $oldPassword
+     */
+    public function setOldPassword(?string $oldPassword): void
+    {
+        $this->oldPassword = $oldPassword;
     }
 
     /**
@@ -33,6 +52,14 @@ class UserPasswordResetData
     public function setNewPassword(?string $newPassword): void
     {
         $this->newPassword = $newPassword;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getOldPassword(): ?string
+    {
+        return $this->oldPassword;
     }
 
     /**
